+++ dantro/plot_creators/__init__.py
@@ -1,18 +1,10 @@
-"""This sub-package implements the dantro PlotCreators
-
-In this module, the following non-abstract plot creators are implemented:
-  - ExternalPlotCreator: imports and calls an external plot script
-  - DeclarativePlotCreator: creates plots using a declarative syntax
-  - ParamSpacePlotCreator: extension of external, making plots from data stored
-    in ParamSpaceGroups more convenient
-  - VegaPlotCreator: interfaces with Altair to provide a Vega-Lite interface
-"""
+"""This sub-package implements non-abstract dantro plot creator classes"""
 
 # Supply the base class
 from dantro.plot_creators.pcr_base import BasePlotCreator
 
 # And all derived PlotCreator classes
-from dantro.plot_creators.pcr_ext import ExternalPlotCreator
+from dantro.plot_creators.pcr_ext import ExternalPlotCreator, is_plot_func
 from dantro.plot_creators.pcr_decl import DeclarativePlotCreator
 from dantro.plot_creators.pcr_psp import UniversePlotCreator
 from dantro.plot_creators.pcr_psp import MultiversePlotCreator