+++ doc/conf.py
@@ -14,7 +14,7 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('.'))
+sys.path.insert(0, os.path.abspath('../dantro'))
 
 
 # -- Project information -----------------------------------------------------
@@ -24,9 +24,9 @@ copyright = '2019, Utopia Developers'
 author = 'Utopia Developers'
 
 # The short X.Y version
-version = '0.6'
+version = '0.7'
 # The full version, including alpha/beta/rc tags
-release = '0.6.0'
+release = '0.7.0rc1'
 
 
 # -- General configuration ---------------------------------------------------